requirements.txt
crispy-bootstrap5==2024.2
django-crispy-forms==2.1